requirements.txt
aiohttp==3.7.4.post0
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
dj-database-url==0.5.0
Django==3.2.6
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
envparse==0.2.0
gunicorn==20.1.0
idna==3.2
multidict==5.1.0
psycopg2==2.9.1
PyJWT==1.7.1
pytz==2021.1
requests==2.26.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.6
whitenoise==5.3.0
yarl==1.6.3
